refactor search components improvement
