+++ src/app/map/components/search-window/search-window.component.html
@@ -1,5 +1,5 @@
 <div class="search-window__searchbar">
-  <search-bar mode="compact" placeholderText="Suchen nach Adressen, Orten, Karten und mehr..."></search-bar>
+  <search-bar mode="compact" [searchConfig]="searchConfig" placeholderText="Suchen nach Adressen, Orten, Karten und mehr..."></search-bar>
   @if (screenMode !== 'mobile') {
     <div class="search-window__results-container">
       <result-groups></result-groups>
