+++ src/app/shared/components/search/search-bar/search-bar.component.html
@@ -7,7 +7,7 @@
     (focusEvent)="handleFocus()"
     [placeholderText]="placeholderText"
     [mode]="mode"
-    [showFilterButton]="screenMode !== 'mobile'"
+    [showFilterButton]="showFilterButton"
     [isAnyFilterActive]="isAnySearchFilterActive"
   ></search-input>
 </div>
