+++ src/app/shared/components/search/search-bar/search-bar.component.scss
