+++ src/app/start-page/components/start-page-search/start-page-search.component.ts
@@ -1,11 +1,8 @@
 import {AfterViewInit, ChangeDetectorRef, Component, Inject, OnDestroy, OnInit, QueryList, ViewChild} from '@angular/core';
-import {MatDialog} from '@angular/material/dialog';
 import {Store} from '@ngrx/store';
 import {combineLatestWith, filter, switchMap, tap} from 'rxjs';
 import {ScreenMode} from 'src/app/shared/types/screen-size.type';
 import {selectScreenMode} from 'src/app/state/app/reducers/app-layout.reducer';
-import {SearchFilterDialogComponent} from '../../../shared/components/search-filter-dialog/search-filter-dialog.component';
-import {PanelClass} from '../../../shared/enums/panel-class.enum';
 import {ConfigService} from '../../../shared/services/config.service';
 import {SearchActions} from '../../../state/app/actions/search.actions';
 import {selectTerm} from '../../../state/app/reducers/search.reducer';
@@ -32,15 +29,14 @@ export class StartPageSearchComponent extends BaseSearchContainerComponent imple
   public searchTerms: string[] = [];
   public activeSearchFilterValues: {groupLabel: string; filterLabel: string}[] = [];
   public screenMode: ScreenMode = 'regular';
+  public readonly searchConfig = this.configService.searchConfig.startPage;
 
-  private readonly searchConfig = this.configService.searchConfig.startPage;
   private readonly screenMode$ = this.store.select(selectScreenMode);
   private readonly searchTerm$ = this.store.select(selectTerm);
   private readonly activeSearchFilterValues$ = this.store.select(selectActiveSearchFilterValues);
 
   constructor(
     private readonly configService: ConfigService,
-    private readonly dialogService: MatDialog,
     @Inject(Store) store: Store,
     @Inject(ChangeDetectorRef) cdr: ChangeDetectorRef,
   ) {
@@ -65,24 +61,10 @@ export class StartPageSearchComponent extends BaseSearchContainerComponent imple
     this.cdr.detectChanges();
   }
 
-  public searchForTerm(term: string) {
-    this.store.dispatch(SearchActions.searchForTerm({term, options: this.searchConfig.searchOptions}));
-  }
-
-  public clearSearchTerm() {
-    this.store.dispatch(SearchActions.clearSearchTerm());
-  }
-
   public deactivateFilter(groupLabel: string, filterLabel: string) {
     this.store.dispatch(SearchActions.setFilterValue({groupLabel, filterLabel, isActive: false}));
   }
 
-  public openFilterMenu() {
-    this.dialogService.open<SearchFilterDialogComponent>(SearchFilterDialogComponent, {
-      panelClass: PanelClass.ApiWrapperDialog,
-      restoreFocus: false,
-    });
-  }
   private initSubscriptions() {
     this.subscriptions.add(
       this.searchTerm$
@@ -109,6 +91,7 @@ export class StartPageSearchComponent extends BaseSearchContainerComponent imple
         )
         .subscribe(),
     );
+
     this.subscriptions.add(this.screenMode$.pipe(tap((screenMode) => (this.screenMode = screenMode))).subscribe());
     this.subscriptions.add(
       this.activeSearchFilterValues$
